widgets: Extract all widgets to separate files
Move each widget class from extension.js into individual files under
widgets/. CPU, Memory, Swap, Network, Disk, GPU, Thermal, Fan,
Frequency, Battery, and Icon each get their own module.

extension.js retains only lifecycle (enable/disable), widget wiring,
and the popup menu. Down from ~2766 lines to ~310 lines.
