+++ system-monitor-next@paradoxxx.zero.gmail.com/widgets/memory.js
@@ -1,0 +1,121 @@
+/* -*- mode: js2; js2-basic-offset: 4; indent-tabs-mode: nil -*- */
+
+import { gettext as _ } from "resource:///org/gnome/shell/extensions/extension.js";
+import Clutter from "gi://Clutter";
+import GTop from "gi://GTop";
+import St from "gi://St";
+import { ElementBase } from '../base.js';
+
+const Mem = class SystemMonitor_Mem extends ElementBase {
+    constructor(extension) {
+        super(extension, {
+            elt: 'memory',
+            elt_short: 'mem',
+            item_name: _('Memory'),
+            color_name: ['program', 'buffer', 'cache']
+        });
+        this.max = 1;
+
+        this.gtop = new GTop.glibtop_mem();
+        this.mem = [0, 0, 0];
+
+        GTop.glibtop_get_mem(this.gtop);
+        this.total = Math.round(this.gtop.total / 1024 / 1024);
+        let threshold = 4 * 1024; // In MiB
+        this.useGiB = false;
+        this._unitConversion = 1024 * 1024;
+        this._decimals = 100;
+        if (this.total > threshold) {
+            this.useGiB = true;
+            this._unitConversion *= 1024 / this._decimals;
+        }
+
+        this.tip_format();
+        this.update();
+    }
+    refresh() {
+        GTop.glibtop_get_mem(this.gtop);
+        if (this.useGiB) {
+            this.mem[0] = Math.round(this.gtop.user / this._unitConversion);
+            this.mem[0] /= this._decimals;
+            this.mem[1] = Math.round(this.gtop.buffer / this._unitConversion);
+            this.mem[1] /= this._decimals;
+            this.mem[2] = Math.round(this.gtop.cached / this._unitConversion);
+            this.mem[2] /= this._decimals;
+            this.total = Math.round(this.gtop.total / this._unitConversion);
+            this.total /= this._decimals;
+        } else {
+            this.mem[0] = Math.round(this.gtop.user / this._unitConversion);
+            this.mem[1] = Math.round(this.gtop.buffer / this._unitConversion);
+            this.mem[2] = Math.round(this.gtop.cached / this._unitConversion);
+            this.total = Math.round(this.gtop.total / this._unitConversion);
+        }
+    }
+    _pad(number) {
+        const Locale = this.extension._Locale;
+        if (this.useGiB) {
+            if (number < 1) {
+                // examples: 0.01, 0.10, 0.88
+                return number.toLocaleString(Locale, {minimumFractionDigits: 2, maximumFractionDigits: 2});
+            }
+            // examples: 5.85, 16.0, 128
+            return number.toLocaleString(Locale, {minimumSignificantDigits: 3, maximumSignificantDigits: 3});
+        }
+
+        return number.toLocaleString(Locale);
+    }
+    _apply() {
+        if (this.total === 0) {
+            this.vals = this.tip_vals = [0, 0, 0];
+        } else {
+            for (let i = 0; i < 3; i++) {
+                this.vals[i] = this.mem[i] / this.total;
+                this.tip_vals[i] = Math.round(this.vals[i] * 100);
+            }
+        }
+        this.text_items[0].text = this.tip_vals[0].toString();
+        this.menu_items[0].text = this.tip_vals[0].toLocaleString(this.extension._Locale);
+        if (this.extension._Style.get('') !== '-compact') {
+            this.menu_items[3].text = this._pad(this.mem[0]) +
+                ' / ' + this._pad(this.total);
+        } else {
+            this.menu_items[3].text = this._pad(this.mem[0]) +
+                '/' + this._pad(this.total);
+        }
+    }
+    create_text_items() {
+        return [
+            new St.Label({
+                text: '',
+                style_class: this.extension._Style.get('sm-status-value'),
+                y_align: Clutter.ActorAlign.CENTER}),
+            new St.Label({
+                text: '%', style_class: this.extension._Style.get('sm-perc-label'),
+                y_align: Clutter.ActorAlign.CENTER})
+        ];
+    }
+    create_menu_items() {
+        let unit = _('MiB');
+        if (this.useGiB) {
+            unit = _('GiB');
+        }
+        return [
+            new St.Label({
+                text: '',
+                style_class: this.extension._Style.get('sm-value')}),
+            new St.Label({
+                text: '%',
+                style_class: this.extension._Style.get('sm-label')}),
+            new St.Label({
+                text: '',
+                style_class: this.extension._Style.get('sm-label')}),
+            new St.Label({
+                text: '',
+                style_class: this.extension._Style.get('sm-value')}),
+            new St.Label({text: unit,
+                style_class: this.extension._Style.get('sm-label')})
+        ];
+    }
+}
+
+export { Mem };
